Move the `render_data_as_html` helper into the library code
