+++ render_data_as_html.py
@@ -4,7 +4,6 @@ import collections
 import datetime
 import json
 import os
-import re
 import shutil
 
 import jinja2
@@ -12,36 +11,8 @@ from jinja2 import Environment, FileSystemLoader, select_autoescape
 from PIL import Image
 import titlecase
 
+from library_lookup.render_data_as_html import display_author_name
 from tint_colors import choose_tint_color_for_file, from_hex
-
-
-def display_author_name(label: str) -> str:
-    """
-    Convert an author name from Spydus into something to display
-    on the page.
-
-    See the tests for examples.
-    """
-    # Remove years of life from the end of the label, e.g.
-    #
-    #     Kawaguchi, Toshikazu, 1971-
-    #     Le Guin, Ursula K., 1929-2018
-    #
-    label = re.sub(r", [0-9]{4}\-(?:[0-9]{4})?$", "", label)
-
-    try:
-        last_name, first_name = label.split(",")
-    except ValueError:
-        return label
-
-    # Remove any trailing descriptions from the first name, e.g.
-    #
-    #     Claire (Journalist)
-    #     Justin (Comic book writer)
-    #
-    first_name = re.sub(r" \([A-Za-z ]+\)$", "", first_name)
-
-    return f"{first_name.strip()} {last_name.strip()}"
 
 
 def rgba(hs: str, opacity: float) -> str:
